+++ rust/src/abc.rs
@@ -9,7 +9,8 @@ use pyo3::exceptions::{
 };
 use pyo3::sync::PyOnceLock;
 use pyo3::types::{
-    PyBool, PyFunction, PyInt, PyIterator, PyList, PySequence, PySet, PyString, PyTuple, PyType,
+    PyBool, PyDict, PyFunction, PyInt, PyIterator, PyList, PyMapping, PyNone, PySequence, PySet,
+    PyString, PyTuple, PyType,
 };
 use pyo3::{BoundObject, IntoPyObjectExt, PyTypeInfo, ffi, intern, prelude::*};
 use tap::prelude::*;
@@ -33,6 +34,8 @@ pub fn abc(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<PyoValuesView>()?;
     m.add_class::<PyoItemsView>()?;
     m.add_class::<PyoMutableSet>()?;
+    m.add_class::<PyoMapping>()?;
+    m.add_class::<PyoMutableMapping>()?;
     Ok(())
 }
 #[pyclass(subclass, frozen, generic, extends=Checkable)]
@@ -1575,20 +1578,85 @@ impl PyoMutableSequence {
             .add_subclass(PyoSequence)
             .add_subclass(Self {})
     }
+
+    #[allow(unused)]
     fn __setitem__<'py>(
         slf: Bound<'py, Self>,
-        _index: Bound<'py, PyAny>,
-        _value: Bound<'py, PyAny>,
+        index: &Bound<'py, PyAny>,
+        value: &Bound<'py, PyAny>,
     ) -> PyResult<()> {
         not_impl_error(slf.as_any(), "PyoMutableSequence", "__setitem__")
     }
-    fn __delitem__<'py>(slf: Bound<'py, Self>, _index: Bound<'py, PyAny>) -> PyResult<()> {
+    #[allow(unused)]
+    fn __delitem__<'py>(slf: Bound<'py, Self>, index: Bound<'py, PyAny>) -> PyResult<()> {
         not_impl_error(slf.as_any(), "PyoMutableSequence", "__delitem__")
     }
+
+    fn __iadd__(slf: Bound<'_, Self>, values: &Bound<'_, PyAny>) -> PyResult<()> {
+        slf.call_method1(intern!(slf.py(), "extend"), (values,))?;
+        Ok(())
+    }
+
+    fn append(slf: Bound<'_, Self>, value: &Bound<'_, PyAny>) -> PyResult<()> {
+        slf.call_method1(intern!(slf.py(), "insert"), (slf.len()?, value))?;
+        Ok(())
+    }
+
+    fn clear(slf: Bound<'_, Self>) -> PyResult<()> {
+        loop {
+            match slf.del_item(0) {
+                Ok(_) => continue,
+                Err(err) => {
+                    if err.is_instance_of::<PyIndexError>(slf.py()) {
+                        return Ok(());
+                    } else {
+                        return Err(err);
+                    }
+                }
+            }
+        }
+    }
+
+    fn extend(slf: Bound<'_, Self>, values: &Bound<'_, PyAny>) -> PyResult<()> {
+        let py = slf.py();
+        if values.is(&slf) {
+            unsafe { slf.cast_into_unchecked::<PySequence>() }.in_place_concat(
+                unsafe { values.cast_unchecked::<PySequence>() }
+                    .to_list()?
+                    .as_sequence(),
+            )?;
+        } else {
+            for v in values.try_iter()? {
+                slf.call_method1(intern!(py, "append"), (v?,))?;
+            }
+        };
+        Ok(())
+    }
+
+    #[pyo3(signature = (index=-1))]
+    fn pop(slf: Bound<'_, Self>, index: isize) -> PyResult<Bound<'_, PyAny>> {
+        let v = slf.get_item(index)?;
+        slf.del_item(index)?;
+        Ok(v)
+    }
+
+    fn remove(slf: Bound<'_, Self>, value: &Bound<'_, PyAny>) -> PyResult<()> {
+        slf.del_item(unsafe { slf.cast_unchecked::<PySequence>() }.index(value)?)
+    }
+
+    fn reverse(slf: Bound<'_, Self>) -> PyResult<()> {
+        let n = slf.len()?;
+        for i in 0..n / 2 {
+            slf.set_item(i, slf.get_item(n - i - 1)?)?;
+            slf.set_item(n - i - 1, slf.get_item(i)?)?;
+        }
+        Ok(())
+    }
+    #[allow(unused)]
     fn insert(
         slf: Bound<'_, Self>,
-        _index: Bound<'_, PyAny>,
-        _value: Bound<'_, PyAny>,
+        index: Bound<'_, PyAny>,
+        value: Bound<'_, PyAny>,
     ) -> PyResult<()> {
         not_impl_error(slf.as_any(), "PyoMutableSequence", "insert")
     }
@@ -2206,5 +2274,287 @@ impl PyoMutableSet {
     fn clear(slf: Bound<'_, Self>) -> PyResult<()> {
         slf.try_iter()?
             .try_for_each(|x| Self::_py_discard(&slf, &x?))
+    }
+}
+#[pyclass(subclass, frozen, generic,  extends=PyoCollection)]
+pub struct PyoMapping;
+#[pymethods]
+impl PyoMapping {
+    #[pyo3(signature = (*_args, **_kwargs))]
+    #[new]
+    fn new(_args: &Args<'_>, _kwargs: Option<&Kwargs<'_>>) -> PyClassInitializer<Self> {
+        PyClassInitializer::from(Checkable)
+            .add_subclass(PyoIterable)
+            .add_subclass(PyoCollection)
+            .add_subclass(Self {})
+    }
+    #[allow(unused)]
+    fn __getitem__<'py>(
+        slf: Bound<'py, Self>,
+        key: Bound<'py, PyAny>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        not_impl_error(slf.as_any(), "PyoMapping", "__getitem__")
+    }
+
+    fn __contains__(slf: Bound<'_, Self>, key: Bound<'_, PyAny>) -> PyResult<bool> {
+        slf.get_item(key).map(|_| true).or_else(|err| {
+            if err.is_instance_of::<PyKeyError>(slf.py()) {
+                Ok(false)
+            } else {
+                Err(err)
+            }
+        })
+    }
+
+    fn __eq__(slf: Bound<'_, Self>, other: Bound<'_, PyAny>) -> PyResult<bool> {
+        let py = slf.py();
+        other
+            .cast::<PyMapping>()
+            .map_err(|_| PyNotImplementedError::new_err(""))
+            .and_then(|other| {
+                slf.call_method0(intern!(py, "items"))?
+                    .pipe_ref(PyDict::from_sequence)?
+                    .eq(other
+                        .call_method0(intern!(py, "items"))?
+                        .pipe_ref(PyDict::from_sequence)?)
+            })
+    }
+
+    fn keys(slf: Bound<'_, Self>) -> PyResult<Bound<'_, PyoKeysView>> {
+        PyoKeysView::type_object(slf.py())
+            .call1((slf,))
+            .map(|x| unsafe { x.cast_into_unchecked::<PyoKeysView>() })
+    }
+
+    fn values(slf: Bound<'_, Self>) -> PyResult<Bound<'_, PyoValuesView>> {
+        PyoValuesView::type_object(slf.py())
+            .call1((slf,))
+            .map(|x| unsafe { x.cast_into_unchecked::<PyoValuesView>() })
+    }
+
+    fn items(slf: Bound<'_, Self>) -> PyResult<Bound<'_, PyoItemsView>> {
+        PyoItemsView::type_object(slf.py())
+            .call1((slf,))
+            .map(|x| unsafe { x.cast_into_unchecked::<PyoItemsView>() })
+    }
+
+    #[pyo3(signature = (key, default=None))]
+    fn get<'py>(
+        slf: Bound<'py, Self>,
+        key: &Bound<'py, PyAny>,
+        default: Option<Bound<'py, PyAny>>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        let py = slf.py();
+        slf.get_item(key).or_else(|err| {
+            if err.is_instance_of::<PyKeyError>(py) {
+                Ok(default.unwrap_or_else(|| PyNone::get(py).into_bound_py_any(py).unwrap()))
+            } else {
+                Err(err)
+            }
+        })
+    }
+
+    fn get_item<'py>(
+        slf: Bound<'py, Self>,
+        key: &Bound<'py, PyAny>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        let py = slf.py();
+        slf.get_item(key)
+            .and_then(|value| PySome::new(value.unbind()).into_bound_py_any(py))
+            .or_else(|err| {
+                if err.is_instance_of::<PyKeyError>(py) {
+                    PyNull::get(py).into_bound_py_any(py)
+                } else {
+                    Err(err)
+                }
+            })
+    }
+}
+
+#[pyclass(subclass, frozen, generic,  extends=PyoMapping)]
+pub struct PyoMutableMapping;
+
+#[pymethods]
+impl PyoMutableMapping {
+    #[pyo3(signature = (*_args, **_kwargs))]
+    #[new]
+    fn new(_args: &Args<'_>, _kwargs: Option<&Kwargs<'_>>) -> PyClassInitializer<Self> {
+        PyClassInitializer::from(Checkable)
+            .add_subclass(PyoIterable)
+            .add_subclass(PyoCollection)
+            .add_subclass(PyoMapping)
+            .add_subclass(Self {})
+    }
+    #[allow(unused)]
+    fn __setitem__(&self, key: &Bound<'_, PyAny>, value: &Bound<'_, PyAny>) -> PyResult<()> {
+        Err(PyKeyError::new_err("KeyError"))
+    }
+    #[allow(unused)]
+    fn __delitem__(&self, key: &Bound<'_, PyAny>) -> PyResult<()> {
+        Err(PyKeyError::new_err("KeyError"))
+    }
+
+    #[pyo3(signature = (key, default=None))]
+    fn pop<'py>(
+        slf: Bound<'py, Self>,
+        key: &Bound<'py, PyAny>,
+        default: Option<Bound<'py, PyAny>>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        match slf.get_item(key) {
+            Ok(value) => {
+                slf.del_item(key)?;
+                Ok(value)
+            }
+            Err(err) => {
+                if err.is_instance_of::<PyKeyError>(slf.py()) {
+                    default.ok_or(err)
+                } else {
+                    Err(err)
+                }
+            }
+        }
+    }
+
+    fn popitem(slf: Bound<'_, Self>) -> PyResult<(Bound<'_, PyAny>, Bound<'_, PyAny>)> {
+        slf.try_iter()?
+            .next()
+            .map(|k| {
+                let key = k?;
+                let value = slf.get_item(&key)?;
+                slf.del_item(&key)?;
+                Ok((key, value))
+            })
+            .unwrap_or_else(|| Err(PyKeyError::new_err("")))
+    }
+
+    fn clear(slf: Bound<'_, Self>) -> PyResult<()> {
+        slf.try_iter()?.try_for_each(|key| slf.del_item(&key?))
+    }
+
+    #[pyo3(signature = (other=None, **kwds))]
+    fn update(
+        slf: Bound<'_, Self>,
+        other: Option<Bound<'_, PyAny>>,
+        kwds: Option<Bound<'_, PyDict>>,
+    ) -> PyResult<()> {
+        other.map(|x| {
+            if x.is_instance_of::<PyMapping>() {
+                x.try_iter()?
+                    .try_for_each(|key| key.and_then(|k| slf.set_item(&k, x.get_item(&k)?)))
+            } else if x.hasattr("keys")? {
+                x.call_method0(intern!(slf.py(), "keys"))
+                    .unwrap()
+                    .try_iter()?
+                    .try_for_each(|key| key.and_then(|k| slf.set_item(&k, x.get_item(&k)?)))
+            } else {
+                x.try_iter()?.try_for_each(|item| {
+                    let tup = item?.cast_into::<PyTuple>()?;
+                    let (key, value) = (tup.get_item(0)?, tup.get_item(1)?);
+                    slf.set_item(&key, &value)
+                })
+            }
+        });
+        kwds.map(|kwds| {
+            kwds.items()
+                .iter()
+                .map(|x| unsafe { x.cast_into_unchecked::<PyTuple>() })
+                .map(|x| unsafe { (x.get_item_unchecked(0), x.get_item_unchecked(1)) })
+                .try_for_each(|(key, value)| slf.set_item(&key, &value))
+        });
+        Ok(())
+    }
+    #[pyo3(signature = (key, default=None))]
+    fn setdefault<'py>(
+        slf: Bound<'py, Self>,
+        key: &Bound<'py, PyAny>,
+        default: Option<Bound<'py, PyAny>>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        let py = slf.py();
+        slf.get_item(key).or_else(|err| {
+            if err.is_instance_of::<PyKeyError>(py) {
+                let default = default
+                    .map(Ok)
+                    .unwrap_or_else(|| PyNone::get(py).into_bound_py_any(py))?;
+                slf.set_item(key, &default)?;
+                Ok(default)
+            } else {
+                Err(err)
+            }
+        })
+    }
+
+    fn insert<'py>(
+        slf: Bound<'py, Self>,
+        key: &Bound<'py, PyAny>,
+        value: &Bound<'py, PyAny>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        let py = slf.py();
+        let previous = slf.get_item(key);
+        slf.set_item(&key, &value)?;
+        previous
+            .map(|x| PySome::new(x.unbind()).into_bound_py_any(py))
+            .unwrap_or_else(|_| PyNull::get(py).into_bound_py_any(py))
+    }
+
+    fn try_insert<'py>(
+        slf: Bound<'py, Self>,
+        key: &Bound<'py, PyAny>,
+        value: Bound<'py, PyAny>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        let py = slf.py();
+        if slf.contains(&key)? {
+            PyoErr::new(
+                PyKeyError::new_err(format!(
+                    "Key {} already exists with value {}.",
+                    key,
+                    slf.get_item(&key)?
+                ))
+                .into_py_any(py)?,
+            )
+            .into_bound_py_any(py)
+        } else {
+            slf.set_item(&key, &value)?;
+            value.unbind().pipe(PyoOk::new).into_bound_py_any(py)
+        }
+    }
+
+    fn remove<'py>(slf: Bound<'py, Self>, key: &Bound<'py, PyAny>) -> PyResult<Bound<'py, PyAny>> {
+        let py = slf.py();
+        slf.get_item(key)
+            .and_then(|value| {
+                slf.del_item(key)?;
+                value.unbind().pipe(PySome::new).into_bound_py_any(py)
+            })
+            .or_else(|err| {
+                if err.is_instance_of::<PyKeyError>(slf.py()) {
+                    PyNull::get(py).into_bound_py_any(py)
+                } else {
+                    Err(err)
+                }
+            })
+    }
+
+    fn remove_entry<'py>(
+        slf: Bound<'py, Self>,
+        key: &Bound<'py, PyAny>,
+    ) -> PyResult<Bound<'py, PyAny>> {
+        let py = slf.py();
+        match slf.get_item(key) {
+            Ok(value) => {
+                slf.del_item(key)?;
+                PyTuple::new(py, [key, &value])?
+                    .into_any()
+                    .unbind()
+                    .pipe(PySome::new)
+                    .into_bound_py_any(py)
+            }
+            Err(err) => {
+                if err.is_instance_of::<PyKeyError>(slf.py()) {
+                    PyNull::get(py).into_bound_py_any(py)
+                } else {
+                    Err(err)
+                }
+            }
+        }
     }
 }